Update documenation
Simple note to self or the next person to use this. All of this really needs to be cleaned up and packaged into one thing. It was built quickly but it's been used several times now. Needs things to be built as classes, more variables (lots of hard coded stuff now), and the whole thing to be packaged as a shell script.